transcribe_meeting/src/requirements.txt
faster-whisper>=0.9.0
pyannote.audio>=2.1.1
torch>=2.0.0
numpy>=1.20.0
ffmpeg-python>=0.2.0
typeguard>=2.13.3
huggingface_hub>=0.19.0
tqdm>=4.65.0
matplotlib>=3.7.0
